Correct documentation of BitMask.topbit
The use of BitMask.topbit was incorrect when BitMasks.shifts was a
non-empty list (in 1.0.0, this affected the max_elt and elements
functions). BitMask.topbit is only used to mean the representation of
the highest valid constant constructor in the sum-type, so the
documentation has been updated to reflect that (no actual code change,
though the name of the value is therefore misleading).