Fix use of Mask.shifts if Mask.lowest <> Mask.one
If Mask.lowest is not Mask.one, then the first element of Mask.shifts
should be the number of zeros before the first 1 in Mask.mask. In this
instance, all the functions using Mask.shifts need to drop the first
element, since by using Mask.lowest that first shift has already been
taken into account.

Test module updated to be more evil to probe this.